requirements.txt
python-engineio==3.13.2 
python-socketio==4.6.0 
Flask-SocketIO==4.3.1 
eventlet

torch
torchvision
numpy
opencv-python
